Update README.md
“BLOOM” is to help people to relieve their mental stress in a much congenial manner by making them engaged in several mind relaxing activities. Mental stress is a social phenomenon which all social groups and all age categories may experience. Hence the beneficiaries and the intended audience of this system will be people of any age category.
BLOOM is developed as a web application and it is supported in mobile platforms as well. This is for the use of the general public and 100% free for anyone to use. Also the users are privileged with sharing their own intentions about the application as feedbacks. All these information is maintained in a database which is located on a web server. Administration is granted permission to access this information for further analysis purposes.
The process is based upon the techniques of machine learning, deep learning and CNN. Initially the user’s emotions will be detected and the system offers the three main features listed below for the user to engage with.
 A virtual text based chat-bot to communicate with
 An emotion based music player which consists of calming music tracks
 Mind relaxing games
System initially recommends a feature for the user depending on the mood and user is free either to continue on the path initiated by the system or choose any feature of preference.

Virtual Chatbot
The chatbot will guide the user suggesting methods to relieve his/her stress by making them engage in several leisure activities. So, the user won’t have to spend time deciding what activity he/she should do.
The friendly text-based conversations with the chatbot can also help to relieve the stress of the user to a great extent.
This can be also seen as a virtual therapy session which will help the user to improve his/her interpersonal skills and eventually learn to open up more about the struggles he/she facing to his/her family members or to closest friends.

Emotion-Based Music Player
The music player will mainly contain of soothing music and ASMR tracks. Following benefits can be gained from listening to relaxing music:
 Stress and Anxiety Relief
 Reduces a person’s perception of pain
 Improves brain functioning
 Helps to sleep better at night
 Lower blood pressure
 Replenish body muscles and limbs
 
System will also assist the user to select a suitable game according to the preference.
The following benefits can be gained from playing anti-stress games:
 Reduce cognitive fatigue
 Reduce stress and anxiety
 Reduce frustration
 Release dopamine, the “happiness hormone”